+++ apps/api/src/services/data-provider/google-sheets/google-sheets.service.ts
@@ -14,6 +14,7 @@ import {
 import { PrismaService } from '@ghostfolio/api/services/prisma/prisma.service';
 import { SymbolProfileService } from '@ghostfolio/api/services/symbol-profile/symbol-profile.service';
 import { DATE_FORMAT, parseDate } from '@ghostfolio/common/helper';
+import { DataProviderInfo } from '@ghostfolio/common/interfaces';
 import { Injectable, Logger } from '@nestjs/common';
 import { DataSource, SymbolProfile } from '@prisma/client';
 import { format } from 'date-fns';
@@ -37,6 +38,12 @@ export class GoogleSheetsService implements DataProviderInterface {
     return {
       dataSource: this.getName(),
       symbol: aSymbol
+    };
+  }
+
+  public getDataProviderInfo(): DataProviderInfo {
+    return {
+      isPremium: false
     };
   }
 
@@ -177,7 +184,11 @@ export class GoogleSheetsService implements DataProviderInterface {
       }
     });
 
-    return { items };
+    return {
+      items: items.map((item) => {
+        return { ...item, dataProviderInfo: this.getDataProviderInfo() };
+      })
+    };
   }
 
   private async getSheet({